rename: 작업 공간 폴더 이름, 구조 컨벤션에 맞게 수정
