+++ src/app/workspace/[id]/components/memo-panel/MemoCard.module.scss
