+++ src/app/workspace/[id]/components/memo-panel/MemoCard.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react'
 
-import IconButton from '@components/Buttons/IconButton/IconButton'
 import { CheckIcon, MoreIcon } from '@components/Icons'
 import Row from '@components/Row/Row'
-import Text from '@components/Text/Text'
+import IconButton from '@components/buttons/IconButton'
+import Text from '@components/text/Text'
 
 import classNames from 'classnames/bind'
