+++ src/app/workspace/[id]/components/memo-panel/MemoPanelBody.tsx
@@ -1,6 +1,6 @@
-import Tab from '@components/Tab/Tab'
+import Tab from '@components/tab/Tab'
 
-import MemoList from '../MemoList/MemoList'
+import MemoList from './MemoList'
 
 export default function MemoPanelBody() {
   return (
