+++ src/app/workspace/[id]/components/panel/PanelBody.module.scss
