+++ src/app/workspace/[id]/components/panel/PanelBody.tsx
