+++ src/app/workspace/[id]/components/panel/PanelContainer.module.scss
