+++ src/app/workspace/[id]/components/panel/PanelHeader.module.scss
